# src/api/openapi.yaml
openapi: 3.1.0
info:
  title: Certificate Manager API
  description: API for managing SSL/TLS certificates
  version: 1.0.0
servers:
  - url: /api
    description: Main API endpoint
...
paths:
...
  /docker/containers:
    get:
      summary: Get all Docker containers
      operationId: getDockerContainers
      tags:
        - docker
      responses:
        "200":
          description: List of Docker containers
          content:
            application/json:
              schema:
                type: object
                properties:
                  dockerAvailable:
                    type: boolean
                    description: Whether Docker is available on the server
                  containers:
                    type: array
                    items:
                      $ref: "#/components/schemas/DockerContainer"
        "503":
          description: Docker service not available
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  dockerAvailable:
                    type: boolean
  /docker/containers/{id}/restart:
    post:
      summary: Restart Docker container
      operationId: restartDockerContainer
      tags:
        - docker
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Container ID or name
      responses:
        "200":
          description: Container restart result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        "503":
          description: Docker service not available
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
...
components:
  schemas:
...
    DockerContainer:
      type: object
      properties:
        id:
          type: string
          description: Full container ID
        shortId:
          type: string
          description: Short container ID (first 12 characters)
        name:
          type: string
          description: Container name
        image:
          type: string
          description: Container image
        status:
          type: string
          description: Container status/state
        created:
          type: number
          description: Creation timestamp
...